requirements.txt
alembic==1.13.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bleach==6.1.0
blinker==1.7.0
bootstrap-flask==2.3.3
cachelib==0.12.0
cffi==1.16.0
click==8.1.7
dnspython==2.6.1
email-validator==2.1.0.post1
flask==3.0.2
flask-admin==1.6.1
flask-cors==4.0.0
flask-login==0.6.3
flask-mailman==1.0.0
flask-migrate==4.0.5
flask-moment==1.0.5
flask-principal==0.4.0
flask-security-too==5.3.3
flask-session==0.6.0
flask-sessions==0.1.5
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
jinja2==3.1.3
mako==1.3.2
markupsafe==2.1.5
nanoid==2.0.0
packaging==23.2
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.1
setuptools==69.1.1
six==1.16.0
sqlalchemy==2.0.27
typing-extensions==4.9.0
webencodings==0.5.1
werkzeug==3.0.1
wtforms==3.1.2
